pelican_nlp-0.3.9/pyproject.toml
[build-system]
requires = ["setuptools>=75.8.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pelican_nlp"
dynamic = ["version"]
description = "Preprocessing and Extraction of Linguistic Information for Computational Analysis"
readme = "README.rst"
requires-python = ">=3.10"
license = "CC-BY-NC-4.0"
authors = [
    {name = "Yves Pauli", email = "yves.pauli@gmail.com"}
]
keywords = ["nlp", "linguistics", "preprocessing", "language-processing", "text-analysis"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic"
]
dependencies = [
    "numpy==2.0.1",
    "pandas==2.2.3",
    "PyYAML>=6.0.2",
    "torch==2.5.1",
    "spacy==3.8.2",
    "transformers==4.49.0",
    "docx2txt>=0.9",
    "striprtf>=0.0.28",
    "chardet>=4.0.0",
    "scikit_learn>=1.6.1",
    "scipy==1.15.2",
    "fasttext-wheel==0.9.2",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    "accelerate==1.4.0",
    "editdistance>=0.8.1",
    "psutil>=6.1.0",
    "tqdm==4.67.1",
    "pytest>=8.3.4",
    "statsmodels>=0.14.4",
    "datasets==3.3.2",
    "huggingface_hub==0.29.2",
    "audiofile>=1.5.1",
    "soundfile>=0.13.1",
    "opensmile>=2.6.0",
    "praat-parselmouth>=0.4.6"
]

[project.urls]
Homepage = "https://github.com/ypauli/pelican_nlp"
Repository = "https://github.com/ypauli/pelican_nlp"
Documentation = "https://github.com/ypauli/pelican_nlp#readme"
"Bug Tracker" = "https://github.com/ypauli/pelican_nlp/issues"

[project.scripts]
pelican-run = "pelican_nlp.cli:main"

[tool.setuptools]
packages = ["pelican_nlp"]
include-package-data = true
package-data = { "pelican_nlp" = ["praat/*.praat"] }

[tool.setuptools.package-dir]
pelican_nlp = "pelican_nlp"

[options]
zip_safe = false

[tool.setuptools.dynamic]
version = {attr = "pelican_nlp._version.__version__"}
